// packages/orbit-components/src/deprecated/Sticky/types.ts
// Type definitions for @kiwicom/orbit-components
// Project: http://github.com/kiwicom/orbit

import * as React from "react";

import * as Common from "../../common/types";

export interface Props extends Common.Globals {
  readonly offset?: number;
  readonly children: React.ReactNode;
}

export interface State {
  sticky: boolean;
  height: number;
  width: number;
  initialTop: number;
  initialWidth: boolean;
}